chore: organize examples and move react to vite
