+++ examples/content-example.ts
