--- examples/react/README.md
@@ -1,3 +1,0 @@
-## React Renderer Example
-
-This example shows how to use the GraphCMS Rich Text Renderer package for React.
