+++ examples/react/package.json
@@ -1,29 +1,23 @@
 {
-  "name": "@graphcms/react-example",
-  "version": "1.0.0",
-  "main": "index.js",
+  "name": "react-example",
   "private": true,
-  "license": "MIT",
+  "version": "0.0.0",
   "scripts": {
-    "start": "parcel index.html",
-    "build": "parcel build index.html",
-    "watch": "parcel watch index.html"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "prismjs": "^1.24.1",
-    "react-app-polyfill": "^1.0.0"
-  },
-  "alias": {
-    "@graphcms/rich-text-types": "../../packages/types",
-    "@graphcms/rich-text-react-renderer": "../../packages/react-renderer",
-    "react": "../../node_modules/react",
-    "react-dom": "../../node_modules/react-dom/profiling",
-    "scheduler/tracing": "../../node_modules/scheduler/tracing-profiling"
+    "prismjs": "^1.27.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
   },
   "devDependencies": {
-    "@types/react": "^17.0.5",
-    "@types/react-dom": "^17.0.3",
-    "parcel": "1.12.3",
-    "typescript": "^4.2.4"
+    "@types/prismjs": "^1.26.0",
+    "@types/react": "^17.0.33",
+    "@types/react-dom": "^17.0.10",
+    "@vitejs/plugin-react": "^1.0.7",
+    "typescript": "^4.5.4",
+    "vite": "^2.8.0"
   }
 }
