+++ examples/svelte/README.md
@@ -1,3 +1,3 @@
-## Svelte Renderer Example
+## Rich Text Renderer with Svelte
 
 This example shows how to use the GraphCMS Rich Text Renderer package with Svelte.
