+++ examples/svelte/src/App.svelte
@@ -1,225 +1,13 @@
 <script lang="ts">
   import { slateToHtml } from '@graphcms/rich-text-html-renderer';
-  import type { RichTextContent, EmbedReferences } from '@graphcms/rich-text-types';
 
-  const content: RichTextContent = {
-    children: [
-      {
-        type: 'heading-two',
-        children: [{ text: 'Awesome new GraphCMS cap!' }],
-      },
-      {
-        type: 'paragraph',
-        children: [
-          { text: 'Sweet black ' },
-          { bold: true, text: 'cap' },
-          { text: ' ' },
-          { text: 'with', underline: true },
-          { text: ' ' },
-          { text: 'embroidered', italic: true },
-          { text: ' ' },
-          { bold: true, text: 'GraphCMS' },
-          { text: ' logo.' },
-        ],
-      },
-      {
-        type: 'bulleted-list',
-        children: [
-          {
-            type: 'list-item',
-            children: [
-              {
-                type: 'list-item-child',
-                children: [{ text: 'Embroided logo' }],
-              },
-            ],
-          },
-          {
-            type: 'list-item',
-            children: [
-              {
-                type: 'list-item-child',
-                children: [{ text: 'Fits well' }],
-              },
-            ],
-          },
-          {
-            type: 'list-item',
-            children: [
-              {
-                type: 'list-item-child',
-                children: [{ text: 'Comes in black' }],
-              },
-            ],
-          },
-          {
-            type: 'list-item',
-            children: [
-              {
-                type: 'list-item-child',
-                children: [{ text: 'Reasonably priced' }],
-              },
-            ],
-          },
-        ],
-      },
-      { type: 'paragraph', children: [{ text: '<Test />', code: true }] },
-      {
-        type: 'code-block',
-        children: [
-          {
-            text: "const teste = 'teste'",
-          },
-        ],
-      },
-      {
-        type: 'code-block',
-        children: [
-          {
-            text: 'const graph = \'cms\'\n\n<?php $echo "Hello" ?>',
-          },
-        ],
-      },
-      {
-        type: 'embed',
-        nodeId: 'cknjbzowggjo90b91kjisy03a',
-        children: [
-          {
-            text: '',
-          },
-        ],
-        nodeType: 'Asset',
-      },
-      {
-        type: 'embed',
-        nodeId: 'ckrus0f14ao760b32mz2dwvgx',
-        children: [
-          {
-            text: '',
-          },
-        ],
-        nodeType: 'Asset',
-      },
-      {
-        type: 'embed',
-        nodeId: 'ckrxv7b74g8il0d782lf66dup',
-        children: [
-          {
-            text: '',
-          },
-        ],
-        nodeType: 'Asset',
-      },
-      {
-        type: 'embed',
-        nodeId: 'ckrxv6otkg6ez0c8743xp9bzs',
-        children: [
-          {
-            text: '',
-          },
-        ],
-        nodeType: 'Asset',
-      },
-      {
-        type: 'embed',
-        nodeId: 'custom_post_id',
-        children: [
-          {
-            text: '',
-          },
-        ],
-        nodeType: 'Post',
-      },
-    ],
-  };
-
-  const references: EmbedReferences = [
-    {
-      id: 'cknjbzowggjo90b91kjisy03a',
-      handle: 'dsQtt0ARqO28baaXbVy9',
-      fileName: 'nkkwzgz0bw6fg6mqzjc1.png',
-      height: 690,
-      width: 880,
-      url: 'https://media.graphcms.com/dsQtt0ARqO28baaXbVy9',
-      mimeType: 'image/png',
-    },
-    {
-      id: 'ckrus0f14ao760b32mz2dwvgx',
-      handle: '7M0lXLdCQfeIDXnT2SVS',
-      fileName: 'file_example_MP4_480_1_5MG.mp4',
-      height: null,
-      width: null,
-      url: 'https://media.graphcms.com/7M0lXLdCQfeIDXnT2SVS',
-      mimeType: 'video/mp4',
-    },
-    {
-      id: 'ckrxv7b74g8il0d782lf66dup',
-      handle: '7VA0p81VQfmZQC9jPB2I',
-      fileName: 'teste.txt',
-      height: null,
-      width: null,
-      url: 'https://media.graphcms.com/7VA0p81VQfmZQC9jPB2I',
-      mimeType: 'text/plain',
-    },
-    {
-      id: 'ckrxv6otkg6ez0c8743xp9bzs',
-      handle: 'HzsAGQyASM2B6B3dHY0n',
-      fileName: 'pdf-test.pdf',
-      height: null,
-      width: null,
-      url: 'https://media.graphcms.com/HzsAGQyASM2B6B3dHY0n',
-      mimeType: 'application/pdf',
-    },
-    {
-      id: 'custom_post_id',
-      title: 'GraphCMS is awesome :rocket:',
-    },
-  ];
+  import { content, references } from '../../content-example';
 
   const html = slateToHtml({ content, references });
 </script>
 
-<main>
-  <h1>Hello Svelte!</h1>
+<div>
+  <h1>Svelte Example</h1>
 
-  <p>{@html html}</p>
-</main>
-
-<style>
-  :root {
-    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen,
-      Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-  }
-
-  main {
-    text-align: center;
-    padding: 1em;
-    margin: 0 auto;
-  }
-
-  h1 {
-    color: #ff3e00;
-    text-transform: uppercase;
-    font-size: 4rem;
-    font-weight: 100;
-    line-height: 1.1;
-    margin: 2rem auto;
-    max-width: 14rem;
-  }
-
-  p {
-    max-width: 14rem;
-    margin: 1rem auto;
-    line-height: 1.35;
-  }
-
-  @media (min-width: 480px) {
-    h1 {
-      max-width: none;
-    }
-
-    p {
-      max-width: none;
-    }
-  }
-</style>
+  <div>{@html html}</div>
+</div>
