+++ examples/vue/README.md
@@ -1,3 +1,3 @@
-## Vue Renderer Example
+## Rich Text Renderer with Vue
 
 This example shows how to use the GraphCMS Rich Text Renderer package with Vue.
